Renamed subnetAllocaiton.js to subnetAllocation.js
